Apply really rudimentary Hunter branding; Add About page; remove placeholder posts
The theme expects a very light theme color, but the Hunter brand color
is dark. As a first step, I took the Hunter brand color and lightened it
so the lightness matches the background that was provided with the
theme. It's not great, but it's more Hunter-ish than the yellow and it
doesn't require re-doing the whole color scheme for the theme.